requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
beautifulsoup4
google-genai
lxml
boto3
PyYAML
fastapi-cors
pydantic